# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-nthroot",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the nth root of `value`.",
        "description": "This service is a thin alias of `srvcs-root`: it does no arithmetic of its\nown and delegates entirely, forwarding `{\"value\", \"n\"}` to `srvcs-root` and\nreturning its `result`. Invalid operands are rejected by `srvcs-root` and\nthe resulting `422` is forwarded unchanged.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "an operand is invalid (forwarded from srvcs-root)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "n"
        ],
        "properties": {
          "n": {
            "type": "object"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "ResultResponse": {
        "type": "object",
        "required": [
          "value",
          "n",
          "result"
        ],
        "properties": {
          "n": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "object"
          }
        }
      }
    }
  }
}